+++ docs/en/changes/changes.md
@@ -64,6 +64,7 @@
 * MQE: Support `&&(and)`, `||(or)` bool operators.
 * OAP self observability: Add JVM heap and direct memory used metrics.
 * OAP self observability: Add watermark circuit break/recover metrics.
+* Add Baseline module for support alarm module query baseline data.
 
 #### UI